+++ composeunstyled-scrollbars/src/commonMain/kotlin/com/composeunstyled/Scrollbars.kt
@@ -75,7 +75,7 @@ class ScrollbarScope internal constructor(
   internal val dragInteraction: MutableState<DragInteraction.Start?>,
   internal val sliderAdapter: SliderAdapter,
   internal val mutableInteractionSource: MutableInteractionSource,
-  internal val scrollAreaState: ScrollAreaState,
+  internal val scrollbarState: ScrollbarState,
 )
 
 sealed class ThumbVisibility {
@@ -89,27 +89,27 @@ sealed class ThumbVisibility {
 
 @Composable
 fun UnstyledVerticalScrollbar(
-  scrollAreaState: ScrollAreaState,
+  scrollbarState: ScrollbarState,
   modifier: Modifier = Modifier,
   enabled: Boolean = true,
   interactionSource: MutableInteractionSource? = null,
   reverseLayout: Boolean = false,
   thumb: @Composable (ScrollbarScope.() -> Unit),
-) = ScrollBar(scrollAreaState, modifier, enabled, interactionSource, reverseLayout, true, thumb)
+) = ScrollBar(scrollbarState, modifier, enabled, interactionSource, reverseLayout, true, thumb)
 
 @Composable
 fun UnstyledHorizontalScrollbar(
-  scrollAreaState: ScrollAreaState,
+  scrollbarState: ScrollbarState,
   modifier: Modifier = Modifier,
   enabled: Boolean = true,
   interactionSource: MutableInteractionSource? = null,
   reverseLayout: Boolean = false,
   thumb: @Composable (ScrollbarScope.() -> Unit),
-) = ScrollBar(scrollAreaState, modifier, enabled, interactionSource, reverseLayout, false, thumb)
+) = ScrollBar(scrollbarState, modifier, enabled, interactionSource, reverseLayout, false, thumb)
 
 @Composable
 private fun ScrollBar(
-  scrollAreaState: ScrollAreaState,
+  scrollbarState: ScrollbarState,
   modifier: Modifier,
   enabled: Boolean = true,
   interactionSource: MutableInteractionSource? = null,
@@ -136,7 +136,7 @@ private fun ScrollBar(
 
   val coroutineScope = rememberCoroutineScope()
   val sliderAdapter = remember(
-    scrollAreaState,
+    scrollbarState,
     containerSize,
     preferredMinThumbSize,
     reverseLayout,
@@ -144,7 +144,7 @@ private fun ScrollBar(
     coroutineScope,
   ) {
     SliderAdapter(
-      scrollAreaState,
+      scrollbarState,
       containerSize,
       preferredMinThumbSize,
       reverseLayout,
@@ -153,12 +153,12 @@ private fun ScrollBar(
     )
   }
 
-  val scrollbarScope = remember(sliderAdapter, resolvedInteractionSource, scrollAreaState) {
+  val scrollbarScope = remember(sliderAdapter, resolvedInteractionSource, scrollbarState) {
     ScrollbarScope(
       dragInteraction,
       sliderAdapter,
       resolvedInteractionSource,
-      scrollAreaState,
+      scrollbarState,
     )
   }
   val scrollThickness = 8.dp.roundToPx()
@@ -217,7 +217,7 @@ fun ScrollbarScope.Thumb(
       var thumbVisibilityJob: Job? by remember { mutableStateOf(null) }
 
       fun shouldKeepThumbVisible(): Boolean {
-        return scrollAreaState.isScrollInProgress || isScrollDragging || isThumbDragging || isTrackHovered
+        return scrollbarState.isScrollInProgress || isScrollDragging || isThumbDragging || isTrackHovered
       }
 
       fun syncThumbVisibility() {
@@ -236,7 +236,7 @@ fun ScrollbarScope.Thumb(
       }
 
       LaunchedEffect(
-        scrollAreaState.isScrollInProgress,
+        scrollbarState.isScrollInProgress,
         isScrollDragging,
         isThumbDragging,
         isTrackHovered,
@@ -245,7 +245,7 @@ fun ScrollbarScope.Thumb(
       }
 
       LaunchedEffect(Unit) {
-        scrollAreaState.interactionSource.interactions
+        scrollbarState.interactionSource.interactions
           .collect { interaction ->
             if (interaction is DragInteraction.Start) {
               isScrollDragging = true
@@ -314,21 +314,21 @@ private val SliderAdapter.thumbPixelRange: IntRange
 private val IntRange.size get() = last + 1 - first
 
 internal class SliderAdapter internal constructor(
-  val adapter: ScrollAreaState,
+  val scrollbarState: ScrollbarState,
   private val trackSize: Int,
   private val minHeight: Float,
   private val reverseLayout: Boolean,
   private val isVertical: Boolean,
   private val coroutineScope: CoroutineScope,
 ) {
-  private val contentSize get() = adapter.contentSize
+  private val contentSize get() = scrollbarState.contentSize
   private val visiblePart: Double
     get() {
       val contentSize = contentSize
       return if (contentSize == 0.0) {
         1.0
       } else {
-        (adapter.viewportSize / contentSize).coerceAtMost(1.0)
+        (scrollbarState.viewportSize / contentSize).coerceAtMost(1.0)
       }
     }
 
@@ -338,12 +338,12 @@ internal class SliderAdapter internal constructor(
   private val scrollScale: Double
     get() {
       val extraScrollbarSpace = trackSize - thumbSize
-      val extraContentSpace = adapter.maxScrollOffset
+      val extraContentSpace = scrollbarState.maxScrollOffset
       return if (extraContentSpace == 0.0) 1.0 else extraScrollbarSpace / extraContentSpace
     }
 
   private val rawPosition: Double
-    get() = scrollScale * adapter.scrollOffset
+    get() = scrollScale * scrollbarState.scrollOffset
 
   val position: Double
     get() = if (reverseLayout) trackSize - thumbSize - rawPosition else rawPosition
@@ -360,7 +360,7 @@ internal class SliderAdapter internal constructor(
     } else {
       value
     }
-    adapter.scrollTo(rawPosition / scrollScale)
+    scrollbarState.scrollTo(rawPosition / scrollScale)
   }
 
   private val dragMutex = Mutex()
@@ -369,7 +369,7 @@ internal class SliderAdapter internal constructor(
     coroutineScope.launch(start = CoroutineStart.UNDISPATCHED) {
       dragMutex.withLock {
         val dragDelta = if (isVertical) offset.y else offset.x
-        val maxScrollPosition = adapter.maxScrollOffset * scrollScale
+        val maxScrollPosition = scrollbarState.maxScrollOffset * scrollScale
         val currentPosition = position
         val targetPosition = (currentPosition + dragDelta + unscrolledDragDistance).coerceIn(
           0.0,
@@ -535,7 +535,7 @@ private class TrackPressScroller(
     offset?.let {
       val currentDirection = directionOfScrollTowards(it)
       if (currentDirection != direction) return
-      with(sliderAdapter.adapter) {
+      with(sliderAdapter.scrollbarState) {
         scrollTo(scrollOffset + currentDirection * viewportSize)
       }
     }