+++ composeunstyled-scrollbars/src/commonTest/kotlin/ScrollBars.test.kt
@@ -52,15 +52,15 @@ class ScrollBarsTest {
   fun always_visible_keeps_the_thumb_visible() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      Box {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState)) {
           repeat(100) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(
@@ -80,15 +80,15 @@ class ScrollBarsTest {
   fun hide_while_idle_starts_with_thumb_out_of_composition() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      Box {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState)) {
           repeat(100) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(
@@ -110,15 +110,15 @@ class ScrollBarsTest {
   fun hide_while_idle_shows_thumb_when_list_is_scrolled() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      Box {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState).testTag("list")) {
           repeat(100) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(
@@ -144,8 +144,8 @@ class ScrollBarsTest {
     // TODO this checks for drags, not scrolls
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      Box {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState).testTag("list")) {
           repeat(100) { index ->
             BasicText("Item $index")
@@ -153,7 +153,7 @@ class ScrollBarsTest {
         }
         Box {
           UnstyledVerticalScrollbar(
-            scrollAreaState = scrollAreaState,
+            scrollbarState = scrollbarState,
             modifier = Modifier.testTag("scrollbar")
               .align(Alignment.CenterEnd),
           ) {
@@ -191,8 +191,8 @@ class ScrollBarsTest {
   fun hide_while_idle_does_not_hide_thumb_while_dragging() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      Box {
         Column(
           modifier = Modifier
             .height(200.dp)
@@ -203,7 +203,7 @@ class ScrollBarsTest {
         }
         Box {
           UnstyledVerticalScrollbar(
-            scrollAreaState = scrollAreaState,
+            scrollbarState = scrollbarState,
             modifier = Modifier
               .testTag("scrollbar")
               .align(Alignment.CenterEnd),
@@ -262,15 +262,15 @@ class ScrollBarsTest {
   fun hide_while_idle_does_not_hide_thumb_while_dragging_the_thumb() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      Box {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState).testTag("list")) {
           repeat(100) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(
@@ -316,15 +316,15 @@ class ScrollBarsTest {
     runComposeUiTest {
       setContent {
         val scrollState = rememberScrollState()
-        val scrollAreaState = rememberScrollAreaState(scrollState)
-        ScrollArea(state = scrollAreaState) {
+        val scrollbarState = rememberScrollbarState(scrollState)
+        Box {
           Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState).testTag("list")) {
             repeat(100) { index ->
               BasicText("Item $index")
             }
           }
           UnstyledVerticalScrollbar(
-            scrollAreaState = scrollAreaState,
+            scrollbarState = scrollbarState,
             modifier = Modifier.testTag("track"),
           ) {
             Thumb(
@@ -376,15 +376,15 @@ class ScrollBarsTest {
   fun vertical_scrollbar_respects_explicit_thumb_height_modifier() = runComposeUiTest {
     setContent {
       val lazyListState = rememberLazyListState()
-      val scrollAreaState = rememberScrollAreaState(lazyListState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(lazyListState)
+      Box {
         LazyColumn(modifier = Modifier.height(200.dp).testTag("list"), state = lazyListState) {
           items((0 until 36_000).toList()) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(